+++ cedar-policy/src/api.rs
@@ -5117,11 +5117,11 @@ mod tpe {
 
     use cedar_policy_core::ast;
     use cedar_policy_core::authorizer::Decision;
-    use cedar_policy_core::tpe;
-    use cedar_policy_core::tpe::batched_evaluator::is_authorized_batched;
-    use cedar_policy_core::tpe::{
-        batched_evaluator::EntityLoader as EntityLoaderInternal, err::BatchedEvalError,
+    use cedar_policy_core::batched_evaluator::is_authorized_batched;
+    use cedar_policy_core::batched_evaluator::{
+        err::BatchedEvalError, EntityLoader as EntityLoaderInternal,
     };
+    use cedar_policy_core::tpe;
     use cedar_policy_core::{
         entities::conformance::EntitySchemaConformanceChecker, extensions::Extensions,
         validator::CoreSchema,
